Refactor ApiServerSection to use ServiceAccountDropdown
