+++ frontend/packages/knative-plugin/src/components/add/event-sources/ApiServerSection.tsx
@@ -1,25 +1,19 @@
 import * as React from 'react';
 import * as _ from 'lodash';
-import * as fuzzy from 'fuzzysearch';
 import { useFormikContext, FormikValues } from 'formik';
 import { FormGroup } from '@patternfly/react-core';
-import { ServiceAccountModel } from '@console/internal/models';
-import { ResourceDropdownField, DropdownField, getFieldId } from '@console/shared';
 import { AsyncComponent } from '@console/internal/components/utils/async';
+import { DropdownField, getFieldId } from '@console/shared';
 import FormSection from '@console/dev-console/src/components/import/section/FormSection';
+import ServiceAccountDropdown from '../../dropdowns/ServiceAccountDropdown';
 
-interface ApiServerSectionProps {
-  namespace: string;
-}
-
-const ApiServerSection: React.FC<ApiServerSectionProps> = ({ namespace }) => {
-  const { values, setFieldValue, setFieldTouched, validateForm } = useFormikContext<FormikValues>();
+const ApiServerSection: React.FC = () => {
+  const { values, setFieldValue } = useFormikContext<FormikValues>();
   const initVal = values?.data?.apiserversource?.resources || [];
   const initialValueResources = !_.isEmpty(initVal)
     ? initVal.map((val) => _.values(val))
     : [['', '']];
   const [nameValue, setNameValue] = React.useState(initialValueResources);
-  const autocompleteFilter = (strText, item): boolean => fuzzy(strText, item?.props?.name);
   const handleNameValuePairs = React.useCallback(
     ({ nameValuePairs }) => {
       const updatedNameValuePairs = _.compact(
@@ -35,25 +29,6 @@ const ApiServerSection: React.FC<ApiServerSectionProps> = ({ namespace }) => {
     },
     [setFieldValue],
   );
-  const onChange = React.useCallback(
-    (selectedValue) => {
-      if (selectedValue) {
-        setFieldTouched('data.apiserversource.serviceAccountName', true);
-        setFieldValue('data.apiserversource.serviceAccountName', selectedValue);
-        validateForm();
-      }
-    },
-    [setFieldValue, setFieldTouched, validateForm],
-  );
-  const resources = [
-    {
-      isList: true,
-      kind: ServiceAccountModel.kind,
-      namespace,
-      prop: ServiceAccountModel.id,
-      optional: true,
-    },
-  ];
   const modeItems = {
     Ref: 'Ref',
     Resource: 'Resource',
@@ -84,18 +59,7 @@ const ApiServerSection: React.FC<ApiServerSectionProps> = ({ namespace }) => {
         title={modeItems.Ref}
         fullWidth
       />
-      <ResourceDropdownField
-        name="data.apiserversource.serviceAccountName"
-        label="Service Account Name"
-        resources={resources}
-        dataSelector={['metadata', 'name']}
-        fullWidth
-        placeholder="Select Service Account Name"
-        onChange={onChange}
-        autocompleteFilter={autocompleteFilter}
-        autoSelect
-        showBadge
-      />
+      <ServiceAccountDropdown name="data.apiserversource.serviceAccountName" />
     </FormSection>
   );
 };
